# Conflicts:
#	iosApp/iosApp.xcodeproj/project.pbxproj
#	iosApp/iosApp/ContentViewModel.swift
#	iosApp/iosApp/Views/Components/Filters/MoreFilterOption.swift
#	iosApp/iosApp/Views/Components/Filters/MoreFilterOptionList.swift
#	iosApp/iosApp/Views/Dashboard/DashboardFilter/DashboardFilterView.swift
#	iosApp/iosApp/Views/Dashboard/DashboardFilter/DashboardFilterViewModel.swift

+++ iosApp/iosApp/Services/DataUploadManager.swift
@@ -21,27 +21,25 @@ class DataUploadManager {
                 print("Fetching Data Bulk...")
                 self.observationDataRepository.allAsBulk { [weak self] dataBulk in
                     if let dataBulk, !dataBulk.dataPoints.isEmpty {
-                        let userDefaults = UserDefaultsRepository()
-                        let networkService = NetworkService(endpointRepository: EndpointRepository(sharedStorageRepository: userDefaults), credentialRepository: CredentialRepository(sharedStorageRepository: userDefaults))
                         print("Sending data to backend...")
-                        
-                        networkService.sendData(data: dataBulk) { [weak self] pair in
-                            if let error = pair.second {
-                                print("Error: \(error)")
-                                self?.currentlyUploading = false
-                                completion(false)
-                            } else if let self, let idSet = pair.first as? Set<String> {
-                                print("Sent data! Deleting local data...")
-                                self.observationDataRepository.deleteAllWithId(idSet: idSet)
-                                print("Deleted data!")
-                                self.currentlyUploading = false
-                                completion(true)
-                            } else {
-                                print("Error!")
-                                self?.currentlyUploading = false
-                                completion(false)
+                        AppDelegate.shared.networkService.iosSendData(data: dataBulk) { weakReference in
+                            if let pair = weakReference.value {
+                                if let error = pair.second {
+                                    print("Error: \(error)")
+                                    self?.currentlyUploading = false
+                                    completion(false)
+                                } else if let self, let idSet = pair.first as? Set<String> {
+                                    print("Sent data! Deleting local data...")
+                                    self.observationDataRepository.deleteAllWithId(idSet: idSet)
+                                    print("Deleted data!")
+                                    self.currentlyUploading = false
+                                    completion(true)
+                                } else {
+                                    print("Error!")
+                                    self?.currentlyUploading = false
+                                    completion(false)
+                                }
                             }
-                            networkService.close()
                         }
                     } else {
                         print("No data to send!")
@@ -54,8 +52,5 @@ class DataUploadManager {
     }
     func close() {
         print("Closed!")
-    }
-    deinit {
-        
     }
 }
